Modify routes to allow users to select individul tracks to download. Include bugfix for issues #9 and #10.
The /spleet route was changed to return a rendered template instead of sending a file to the user.
This template will show the individual files the user can download via a /download route which
will be added later. This change also fixes the persistant loading spinner issue in the UI. Also added
user input validation to ensure a request is not sent with no file selected.